\name{matrixNULL-class}
\docType{class}
\alias{matrixNULL-class}

\title{Virtual Class "matrixNULL" - Simple Class for matrix or NULL}

\description{The class \code{"matrixNULL"} is a simple class union 
  (\code{\link{setClassUnion}}) of \code{"matrix"} and \code{"NULL"}.
}

\section{Objects from the Class}{Since it is a virtual Class, no objects
  may be created from it.}

\examples{
showClass("matrixNULL")
}

\keyword{classes}
