\name{parsimony_monitor}
\alias{parsimony_monitor}

\title{Functions for monitoring GA-PARSIMONY algorithm evolution}

\description{Functions to print summary statistics of fitness values at each iteration of a GA search.}

\usage{
parsimony_monitor(object, digits = getOption("digits"), \dots)
}

\arguments{
  \item{object}{An object of class \code{"ga_parsimony"}, usually resulting from a call to function \code{\link{ga_parsimony}}.}
  \item{digits}{minimal number of significant digits.}
  \item{\dots}{Further arguments passed to or from other methods.}
}

%\details{}
\value{These functions print a summary of current GA-PARSIMONY step on the console.\cr
}

%\references{}

\author{Francisco Javier Martinez de Pison. \email{fjmartin@unirioja.es}. EDMANS Group. \url{https://edmans.webs.com/}}

%\seealso{\code{\link{ga}}}

%\note{}

%\examples{}

%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
